working implementation with controller methods for own resources

+++ src/controllers/usersController.js
@@ -19,6 +19,22 @@ async function all (req, res) {
   }
 }
 
+async function own (req, res) {
+  try {
+    const result = await User.query().where({ id: res.locals.user.username })
+    const ownUser = result.length ? result[0] : null
+    if (!ownUser) {
+      throw new Error('users own record does not exist: ' + res.locals.user.username)
+    }    
+    res.status(HttpStatus.OK).send(ownUser)
+  } catch (error) {
+    logger.error(error.message)
+    res.status(HttpStatus.INTERNAL_SERVER_ERROR).send({
+      error: HttpStatus.getStatusText(HttpStatus.INTERNAL_SERVER_ERROR)
+    })
+  }
+}
+
 async function findOne (req, res) {
   try {
     const result = await User.query().where({ id: req.params.id })
@@ -80,6 +96,24 @@ async function updateOne (req, res) {
   }
 }
 
+async function updateOwn (req, res) {
+  let user
+  try {
+    const userExists = await User.query().where({ id: res.locals.user.username })
+    if (!userExists.length) {
+      throw new Error('users own record does not exist: ' + res.locals.user.username)
+    } else {
+      user = await User.query().patchAndFetchById(res.locals.user.username, req.body)
+    }
+    res.status(HttpStatus.OK).send(user)
+  } catch (error) {
+    logger.error(error.message)
+    res.status(HttpStatus.INTERNAL_SERVER_ERROR).send({
+      error: HttpStatus.getStatusText(HttpStatus.INTERNAL_SERVER_ERROR)
+    })
+  }
+}
+
 async function deleteOne (req, res) {
   const trx = await PostgresDB.startTransaction()
   try {
@@ -107,4 +141,31 @@ async function deleteOne (req, res) {
   }
 }
 
-module.exports = { all, findOne, addOne, updateOne, deleteOne }
+async function deleteOwn (req, res) {
+  const trx = await PostgresDB.startTransaction()
+  try {
+    const user = await User.query(trx)
+      .where({ id: res.locals.user.username })
+      .first()
+    if (user instanceof User) {
+      await user.$relatedQuery('organizations').delete()
+      await user.$relatedQuery('notifications').delete()
+      await user.$relatedQuery('offers').delete()
+      await User.query(trx)
+        .delete()
+        .where({ id: res.locals.user.username })
+    } else {
+      throw new Error('No user.id: ' + res.locals.user.username)
+    }
+    await trx.commit()
+    res.status(HttpStatus.OK).send()
+  } catch (error) {
+    logger.error(error.message)
+    await trx.rollback()
+    res.status(HttpStatus.INTERNAL_SERVER_ERROR).send({
+      error: HttpStatus.getStatusText(HttpStatus.INTERNAL_SERVER_ERROR)
+    })
+  }
+}
+
+module.exports = { all, own, findOne, addOne, updateOne, updateOwn, deleteOne, deleteOwn }
